+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # tiltIndicator 0.0.0.9029
 
-* In PCTR, if a copmany matches no input, all shares are now `NA` (#205).
+* In PCTR, if a company matches no input, all shares are now `NA` (#205).
 
 # tiltIndicator 0.0.0.9028